wu-move-from-tensor: Move update_weights_from_tensor and helpers onto WeightUpdater
Refactor chain ID: wu-move-from-tensor
