+++ test/registered/rl/test_update_weights_from_tensor.py
@@ -97,7 +97,7 @@ class TestUpdateWeightsFromTensor(CustomTestCase):
 
     def test_update_weights_from_tensor_load_format_custom(self):
         custom_loader_name = (
-            "sglang.srt.model_executor.model_runner._model_load_weights_direct"
+            "sglang.srt.model_executor.weight_updater._model_load_weights_direct"
         )
         engine = sgl.Engine(
             model_path=DEFAULT_SMALL_MODEL_NAME_FOR_TEST,
